passed user State on props
